Resolve config typing errors
